+++ ngmix/em/__init__.py
@@ -1,0 +1,5 @@
+# flake8: noqa
+
+from . import em
+from .em import *
+from . import em_nb
